ENCDataVault: Split into two formats
Splits the ENCDataVault format in two format structs and splits files.
Exposes the PBKDF2-HMAC-SHA256 iteration count as a tunable cost.
Updates the extraction script with the new tag name.
Update hash extraction according to Miroslav Dimitrov remarks
